@@ -1,11 +1,7 @@
-'use strict';
+import { spawnSync } from 'child_process';
 
-const { spawnSync } = require('child_process');
+import config from './../../config.js';
 
-const config = require('./../../config');
-
-function open(notes) {
+export function open(notes) {
   spawnSync(config.EDITOR, notes);
 }
-
-module.exports = open;
