@@ -1,7 +1,3 @@
-'use strict';
-
-function escapeFilename(str) {
+export function escapeFilename(str) {
   return str.replace(/(["\s'$`\\])/g,'\\$1');
 }
-
-module.exports = escapeFilename;
