@@ -1,17 +1,18 @@
 #!/usr/bin/env node
-
 import { Command } from 'commander';
-import fs from 'node:fs';
+import { readFile } from 'node:fs/promises';
 
-const pkg = require('./../package.json');
-const package_ = JSON.parse(fs.readFileSync('package.json'));
+let package_ = JSON.parse(
+  await readFile(
+    new URL('./../package.json', import.meta.url)
+  )
+);
 
 const program = new Command();
 program
   .version(package_.version)
+  .description(`${package_.description}`)
   .command('new', 'create new note')
   .command('list', 'list notes', { isDefault: true }).alias('ls')
   .command('tags', 'list tags')
-  .description(`${pkg.description}`);
-
-program.parse(process.argv);
+  .parse();
